pyproject.toml
[project]
name = "generativeaiwithpython_material"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "absl-py>=2.2.2",
    "datasets>=3.5.0",
    "google-adk>=0.2.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.23",
    "langchain-community>=0.3.21",
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.14",
    "llama-index>=0.12.31",
    "llama-index-indices-managed-llama-cloud>=0.6.11",
    "llama-index-tools-tavily-research>=0.3.0",
    "llama-index-utils-workflow>=0.3.1",
    "python-dotenv>=1.1.0",
    "transformers>=4.51.3",
    "torch==2.6.0",
    "torchvision",
    "torchaudio",
    "langchain-google-genai>=2.1.3",
    "ollama>=0.4.8",
    "langchain-ollama>=0.3.2",
    "sentence-transformers>=4.1.0",
    "faiss-cpu>=1.11.0",
    "langchain-faiss>=0.1.1",
    "seaborn>=0.13.2",
    "langchain-experimental>=0.3.4",
    "langchain-chroma>=0.2.3",
    "streamlit>=1.45.1",
    "langchain-pymupdf4llm>=0.4.1",
    "openai-agents[litellm]>=0.1.0",
    "pydantic>=2.11.3",
    "anthropic>=0.56.0",
]
